Added ThemeCycle MiniTweak with QuickStart Graphic
ThemeCycle is an automation tweak for Windows that allows your PC to automatically switch betwen dark and light modes based on the time of day. This is a popular feature available on other platforms such as iOS, and even Microsoft's Xbox operating system. Now, with this tweak, we can have it on Windows too while we wait for Microsoft to eventually get around to implimenting it...if ever.